+++ zorp-frontend/src/app/zorp/study/encryption-key-cid/page.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { useCallback, useId, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useReadContract } from 'wagmi';
 import { useContracts } from '@/contexts/Contracts';
+import ZorpStudyAddressInput from '@/components/contracts/ZorpStudyAddressInput';
 import ThemeSwitch from '@/components/features/ThemeSwitch';
 import * as config from '@/lib/constants/wagmiConfig';
 
@@ -10,8 +11,6 @@ export default function ZorpStudyReadEncryptionKeyCid() {
 	const addressStudyAnvil = config.anvil.contracts.IZorpStudy[31337].address;
 
 	const [addressStudy, setAddressStudy] = useState<`0x${string}`>(addressStudyAnvil);
-
-	const addressStudyId = useId();
 
 	const { IZorpStudy } = useContracts();
 
@@ -30,10 +29,6 @@ export default function ZorpStudyReadEncryptionKeyCid() {
 			enabled: false,
 		},
 	});
-
-	const handleChangeStudyAddress = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-		setAddressStudy(event.target.value as `0x${string}`);
-	}, [ setAddressStudy ]);
 
 	const handleOnClick = useCallback(async () => {
 		if (!enabled) {
@@ -55,12 +50,9 @@ export default function ZorpStudyReadEncryptionKeyCid() {
 				<ThemeSwitch />
 			</div>
 
-			<label htmlFor={addressStudyId}>ZORP Study Address:</label>
-			<input
-				id={addressStudyId}
-				value={addressStudy}
-				onChange={handleChangeStudyAddress}
+			<ZorpStudyAddressInput
 				disabled={isFetching}
+				setState={setAddressStudy}
 			/>
 
 			<button
